backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt==2.8.1
python-multipart==0.0.6
aiofiles==23.2.1
aiodns==3.1.1
cchardet==2.1.7
PyYAML==6.0.1
requests==2.31.0
openai==1.3.9
google-cloud-language==2.11.1
google-generativeai==0.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
httpx==0.25.2
stripe==7.4.0
reportlab==4.0.7
cryptography==41.0.7
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.2
python-dateutil==2.8.2
pytz==2023.3
pydantic-extra-types==2.4.0
